+++ API/EnterprisesController.cs
@@ -10,6 +10,7 @@ using CarPark.APIModels;
 using Microsoft.AspNetCore.Authorization;
 using System.Data;
 using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
 
 namespace CarPark.API
 {
@@ -18,13 +19,21 @@ namespace CarPark.API
     [Authorize]
     public class EnterprisesController : ControllerBase
     {
-        private readonly ApplicationDbContext _context;
-        readonly UserManager<ApplicationUser> _userManager;
+        readonly ApplicationDbContext _context;
+        readonly int _userId;
 
-        public EnterprisesController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
+        public EnterprisesController
+        (
+            ApplicationDbContext context,
+            IHttpContextAccessor contextAccessor,
+            UserManager<ApplicationUser> userManager
+        )
         {
             _context = context;
-            _userManager = userManager;
+
+            var user = contextAccessor.HttpContext?.User;
+            string? userId = user is not null ? userManager.GetUserId(user) : null;
+            _userId = Int32.Parse(userId ?? "");
         }
 
         // GET: api/Enterprises
@@ -34,7 +43,7 @@ namespace CarPark.API
             if (_context.Enterprises == null)
                 return NotFound();
 
-            var userEnterprises = await Enterprise.GetUserEnterprises(_context, _userManager, User);
+            var userEnterprises = Enterprise.GetUserEnterprises(_context, User, _userId);
 
             return await userEnterprises
                 .Include(e => e.Drivers)
@@ -51,9 +60,7 @@ namespace CarPark.API
             if (_context.Enterprises == null)
                 return NotFound();
 
-            var userEnterprises = await Enterprise.GetUserEnterprises(_context, _userManager, User);
-
-            var enterprise = await userEnterprises
+            var enterprise = await _context.Enterprises
                 .Include(e => e.Drivers)
                     .ThenInclude(d => d.ActiveVehicle)
                 .Include(e => e.Vehicles)
@@ -61,6 +68,9 @@ namespace CarPark.API
 
             if (enterprise == null)
                 return NotFound();
+
+            if (!(await Enterprise.CheckAccess(id, _context, User, _userId)))
+                return StatusCode(StatusCodes.Status403Forbidden);
 
             return new EnterpriseAPIModel(enterprise);
         }
@@ -72,6 +82,9 @@ namespace CarPark.API
         {
             if (id != enterprise.Id)
                 return BadRequest();
+
+            if (!(await Enterprise.CheckAccess(id, _context, User, _userId)))
+                return StatusCode(StatusCodes.Status403Forbidden);
 
             _context.Entry(enterprise).State = EntityState.Modified;
 
@@ -98,7 +111,7 @@ namespace CarPark.API
             if (_context.Enterprises == null)
                 return Problem("Entity set 'AppDbContext.Enterprises'  is null.");
 
-            await Enterprise.SaveAndBindWithManager(_context, _userManager, User, enterprise);
+            await enterprise.SaveAndBindWithManager(_context, User, _userId);
             return CreatedAtAction("GetEnterprise", new { id = enterprise.Id }, new EnterpriseAPIModel(enterprise));
         }
 
@@ -114,12 +127,14 @@ namespace CarPark.API
             if (enterprise == null)
                 return NotFound();
 
+            if (!(await Enterprise.CheckAccess(id, _context, User, _userId)))
+                return StatusCode(StatusCodes.Status403Forbidden);
+
             _context.Enterprises.Remove(enterprise);
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
-
         private bool EnterpriseExists(int id)
         {
             return (_context.Enterprises?.Any(e => e.Id == id)).GetValueOrDefault();
